Restructure folders to follow MVVM with Clean Architecture
