+++ app/src/main/java/acn/assessment/weatherapp/MainActivity.kt
@@ -1,5 +1,8 @@
 package acn.assessment.weatherapp
 
+import acn.assessment.weatherapp.LocationPermissionDetails
+import acn.assessment.weatherapp.LocationPermissionNotAvailable
+import acn.assessment.weatherapp.R
 import android.Manifest
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -25,10 +28,11 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import acn.assessment.weatherapp.service.dto.CurrentWeather
-import acn.assessment.weatherapp.service.dto.FullWeather
+import acn.assessment.weatherapp.service.data.dto.CurrentWeather
+import acn.assessment.weatherapp.service.data.dto.FullWeather
 import acn.assessment.weatherapp.ui.theme.ThemeState
 import acn.assessment.weatherapp.ui.theme.WeatherAppTheme
+import acn.assessment.weatherapp.viewmodels.MainViewModel
 import androidx.compose.foundation.background
 import androidx.compose.material.*
 import androidx.compose.runtime.*
@@ -154,7 +158,6 @@ fun WeatherSummary(weather: CurrentWeather) {
 @Composable
 fun TemperatureSummary(weather: CurrentWeather) {
     Surface {
-
         Row(
             Modifier
                 .fillMaxWidth()
